adds media file support for content import
